Consolidate other party menu source files
