+++ include/applications/party_menu/sprites.h
