+++ platinum.us/main.lsf
@@ -306,10 +306,10 @@ Static main
 	Object main.nef.p/src_unk_0207DFAC.c.o
 	Object main.nef.p/src_unk_0207E060.c.o
 	Object main.nef.p/src_applications_party_menu_main.c.o
-	Object main.nef.p/src_unk_020819DC.c.o
-	Object main.nef.p/src_unk_02082C2C.c.o
-	Object main.nef.p/src_unk_02083370.c.o
-	Object main.nef.p/src_unk_02084B70.c.o
+	Object main.nef.p/src_applications_party_menu_windows.c.o
+	Object main.nef.p/src_applications_party_menu_sprites.c.o
+	Object main.nef.p/src_applications_party_menu_unk_02083370.c.o
+	Object main.nef.p/src_applications_party_menu_unk_02084B70.c.o
 	Object main.nef.p/src_applications_naming_screen.c.o
 	Object main.nef.p/src_unk_020890F4.c.o
 	Object main.nef.p/src_unk_02089604.c.o
