+++ src/meson.build
@@ -289,10 +289,10 @@ pokeplatinum_c = files(
     'unk_0207DFAC.c',
     'unk_0207E060.c',
     'applications/party_menu/main.c',
-    'unk_020819DC.c',
-    'unk_02082C2C.c',
-    'unk_02083370.c',
-    'unk_02084B70.c',
+    'applications/party_menu/windows.c',
+    'applications/party_menu/sprites.c',
+    'applications/party_menu/unk_02083370.c',
+    'applications/party_menu/unk_02084B70.c',
     'applications/naming_screen.c',
     'unk_020890F4.c',
     'unk_02089604.c',
